Filter files uploaded for supported ones
The current behavior of qvis is to reject all files uploaded
together when one of them cannot be accepted because of an
invalid filename.

This patch modifies that behavior so that only those files
that do not have an acceptable file extension are rejected
and the others are properly parsed.